requirements.txt
Django==3.2.8
django-environ==0.4.5  # For managing environment variables
gunicorn==20.1.0  # A production-ready WSGI server
django-model-utils==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==3.13.0
